fix: changes due to PR
